Move Text lookup repos to Persistence (F18).
KnownTermsRepository and provider factories now live in Persistence/Lookups; Text drops Persistence.Abstractions reference and AddEliminationTerms moves to Persistence.Extensions.
