+++ Class-Libraries/RedditPodcastPoster.Persistence/Lookups/EliminationTermsProviderFactory.cs
@@ -1,0 +1,19 @@
+using Microsoft.Extensions.Logging;
+using RedditPodcastPoster.Persistence.Abstractions;
+using RedditPodcastPoster.Text.EliminationTerms;
+using EliminationTermsModel = RedditPodcastPoster.Models.EliminationTerms;
+
+namespace RedditPodcastPoster.Persistence.Lookups;
+
+public class EliminationTermsProviderFactory(
+    IEliminationTermsRepository eliminationTermsRepository,
+    ILogger<EliminationTermsProviderFactory> logger)
+    : IEliminationTermsProviderFactory
+{
+    public async Task<IEliminationTermsProvider> Create()
+    {
+        logger.LogInformation($"{nameof(Create)} - Creating {nameof(EliminationTermsProvider)}");
+        var terms = await eliminationTermsRepository.Get();
+        return new EliminationTermsProvider(terms ?? new EliminationTermsModel());
+    }
+}
