--- Class-Libraries/RedditPodcastPoster.Text/EliminationTerms/EliminationTermsProviderFactory.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-using RedditPodcastPoster.Persistence.Abstractions;
-
-namespace RedditPodcastPoster.Text.EliminationTerms;
-
-public class EliminationTermsProviderFactory(
-    ILookupRepository lookupRepository,
-    ILogger<EliminationTermsProviderFactory> logger)
-    : IEliminationTermsProviderFactory
-{
-    public async Task<IEliminationTermsProvider> Create()
-    {
-        logger.LogInformation($"{nameof(Create)} - Creating {nameof(EliminationTermsProvider)}");
-        var terms = await lookupRepository.GetEliminationTerms();
-        return new EliminationTermsProvider(terms ?? new RedditPodcastPoster.Models.EliminationTerms());
-    }
-}
