+++ Class-Libraries/RedditPodcastPoster.Text/Extensions/ServiceCollectionExtensions.cs
@@ -1,7 +1,4 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using RedditPodcastPoster.DependencyInjection;
-using RedditPodcastPoster.Text.EliminationTerms;
-using RedditPodcastPoster.Text.KnownTerms;
 
 namespace RedditPodcastPoster.Text.Extensions;
 
@@ -9,23 +6,16 @@ public static class ServiceCollectionExtensions
 {
     extension(IServiceCollection services)
     {
+        /// <summary>
+        /// Text sanitisation services. Requires <c>AddRepositories()</c> (Persistence) for
+        /// <see cref="KnownTerms.IKnownTermsProvider"/> resolution via Cosmos lookup repos.
+        /// </summary>
         public IServiceCollection AddTextSanitiser()
         {
             return services
                 .AddSingleton<ITextSanitiser, TextSanitiser>()
                 .AddSingleton<IHtmlSanitiser, HtmlSanitiser>()
-                .AddSingleton<IKnownTermsProviderFactory, KnownTermsProviderFactory>()
-                .AddSingleton<IAsyncInstance<IKnownTermsProvider>>(s =>
-                    new AsyncInstance<IKnownTermsProvider>(s.GetService<IKnownTermsProviderFactory>()!))
                 .AddSingleton<IHashTagEnricher, HashTagEnricher>();
-        }
-
-        public IServiceCollection AddEliminationTerms()
-        {
-            return services
-                .AddSingleton<IEliminationTermsProviderFactory, EliminationTermsProviderFactory>()
-                .AddSingleton<IAsyncInstance<IEliminationTermsProvider>>(s =>
-                    new AsyncInstance<IEliminationTermsProvider>(s.GetService<IEliminationTermsProviderFactory>()!));
         }
     }
 }
